+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Added
 - Messages being sent using TRP3's NPC speech now persist the NPC speech formatting through all messages.
 	- You can toggle this behavior in the Chattery settings (found in the Addon Compartment or through `/chattery`)
+- Added hints to settings in the settings UI
 
 ## Fixed
 - Fixed an issue sending certain types of chat messages while in a garrison.